Add onMessage listener synchronously
Unsure if it really matters for `runtime.onMessage` but in general
Chrome docs warns that listeners must be registered on extension loading
synchronously. Otherwise a problem could happen in extension is
unloaded. Though this extension use messages only in handlers
for other events, so likely the change is not necessary.