Restore settings context menu entry for browser action
Add a direct way to open add-on settings from the extension entry
in Firefox toolbar. Chrome provides "options" menu entry for this
purpose. In Firefox a similar "Manage Extension" item leads
to the add-on manager and extension settings are hidden
in the "..." menu. While `options_ui` page was not opened
in a separate tab, the extension manager had the easily discoverable
"Preferences" tab.

This commit effectively reverts

    588429ee9d 2021-12-06 Remove "Settings" from browser action context menu

for Firefox.

Partially update `README.org` and the help page.